Changed testing suite from Jest to Mocha
